% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auroc_aupr.R
\name{auroc_aupr}
\alias{auroc_aupr}
\title{Function to fetch score of AUROC and AUPR}
\usage{
auroc_aupr(scores, labels, mode = "both")
}
\arguments{
\item{labels}{A numeric, character, logical, or factor dataset of observed labels}

\item{mode}{A string that specifies the types of evaluation measures that the evalmod function calculates, roc, pr, both.
the default is both.}

\item{score}{A numeric dataset of predicted scores, a vector}
}
\value{
value of computed scores of roc or prc.
}
\description{
This function computes the value of AUROC and AUPR at the same time
}
\examples{
x <- c(0.1, 0.5, 0.6, 0.2)
y <- factor(c('yes','no','yes','no'))
auroc_aupr(scores = x, labels = y)
}
\seealso{
\code{\link{[precrec]evalmod}}
}
