% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comb_matrix_to_df.R
\name{comb_matrix_to_df}
\alias{comb_matrix_to_df}
\title{Function to convert named matrix to colname-rowname dataframe}
\usage{
comb_matrix_to_df(input.matrix = matrix(), colnames.first = TRUE,
  new.colnames = c("c", "r", "value"))
}
\arguments{
\item{input.matrix}{Matrix contains names and values}

\item{colnames.first}{Whether in the return dataframe, the colnames of matrix should be the first col}

\item{new.colnames}{The colnames of the return dataframe}
}
\value{
A dataframe contains colnames and rownames of the input matrix, as well as the value
}
\description{
This function helps you to convert your matrix into dataframe if your matrix has colnames and rownames with value in it
}
\examples{
\dontrun{
mx <- matrix(1:6, ncol=2)
colnames(mx) <- c("v1", "v2")
rownames(mx) <- c('r1', 'r2', 'r3')
print(comb_matrix_to_df(mx))
}
}
