% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_to_comb_matrix.R
\name{df_to_comb_matrix}
\alias{df_to_comb_matrix}
\title{Function to convert sepcific columns of a dataframe to a named matrix with value}
\usage{
df_to_comb_matrix(input.df = data.frame(), col.margin = 1,
  row.margin = 2, value.margin = 3, fill = NA)
}
\arguments{
\item{input.df}{The dataframe contains needed columns}

\item{col.margin}{The index of inputed dataframe's column to be the colname}

\item{row.margin}{The index of inputed dataframe's column to be the rowname}

\item{value.margin}{The index of inputed dataframe's column to be the value}

\item{fill}{The filled in value when no value of a pair is noticed}
}
\value{
A names matrix with values
}
\description{
Function to convert sepcific columns of a dataframe to a named matrix with value
}
\note{
The value of fill will be filled in if no value of a pair is noticed
}
\examples{
\dontrun{
  A <- data.frame(x = c('a', 'b', 'c'), y = c('A', 'B', 'c'), v = c(3:5))
  df_to_comb_matrix(A, col.margin = 2, row.margin = 1, value.margin = 3)
}
}
