% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfbmd-package.R
\docType{package}
\name{dfbmd}
\alias{dfbmd}
\title{Documented Functions of Bioinformatics during Master Degree}
\description{
Several functions and algorithms related to bioinformatics during my master degree
}
