% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagonal_zero.R
\name{diagonal_zero}
\alias{diagonal_zero}
\title{Make the element of a matrix's diagonal line into zero}
\usage{
diagonal_zero(input = matrix())
}
\arguments{
\item{input}{A matrix has the same size of columns and rows}
}
\value{
A matrix with all element of diagonal line  zero
}
\description{
Make the element of a matrix's diagonal line into zero
}
\examples{
\dontrun{
A <- matrix(1:100, nrow=10)
A <- diagonal_zero(A)
}
}
