% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downladGEOData.R
\name{downloadGEOData}
\alias{downloadGEOData}
\title{Function to get GEO data easily}
\usage{
downloadGEOData(GEO = "", destdir = ".")
}
\arguments{
\item{GEO}{A character string representing a GEO object for download and parsing}

\item{destdir}{The destination directory for any downloads}
}
\value{
An object of the appropriate class (GDS, GPL, GSM, or GSE) is returned
}
\description{
This function writes out two csv files, one contains the data matrix and the other contains the phenotype information of each sample
}
\examples{
downloadGEOData(GEO='GSE39582')
}
\seealso{
\code{\link[GEOquery]{getGEO}}
}
