% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{failthful}
\alias{failthful}
\title{Failthful dataset}
\format{
A data frame contains 272 observations and 2 variables
\describe{
   \item{eruptions}{eruptions, numerical}
   \item{waiting}{waiting, numerical}
}
}
\usage{
failthful
}
\description{
Failthful dataset
}
\keyword{datasets}
