% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forceGraph.R
\name{forceGraph}
\alias{forceGraph}
\title{Create a force-directed graph widget}
\usage{
forceGraph(message, width = NULL, height = NULL)
}
\arguments{
\item{message}{the message of the widget}

\item{width}{the width of the widget}

\item{height}{the height of the widget}
}
\description{
This function creates a force-directed graph using \pkg{htmlwidgets}, which can be rendered
in R Markdown documents and shiny applications.
The rendering is conducted using a variant of the Sigma which is a JavaScript library dedicated
to graph drawing. The force-graph layout algorithm is ForceAtlas2
[http://www.plosone.org/article/info%3Adoi%2F10.1371%2Fjournal.pone.0098679] and implemented by
Mathieu Jacomy [https://github.com/jacomyma], Guillaume Plique [https://github.com/Yomguithereal]
and Sébastien Heymann. [https://github.com/sheymann].
}
