% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{iris}
\alias{iris}
\title{iris data set}
\format{
A data frame with 150 rows and 5 variables:
\describe{
  \item{Sepal.Length}{length of sepal, numerical}
  \item{Sepal.Width}{width of sepal, numerical}
  \item{Petal.Length}{length of petal, numerical}
  \item{Petal.Width}{width of petal, numerical}
  \item{Species}{name of species, factor}
}
}
\usage{
iris
}
\description{
iris data set
}
\examples{
data(iris)
}
\keyword{datasets}
