% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KmeansR.R
\name{KmeansR}
\alias{KmeansR}
\title{My own kmeans with raw R codes}
\usage{
KmeansR(dataset = NA, k = NA)
}
\arguments{
\item{dataset}{Dataframe of your input data}

\item{k}{Number of clusters}
}
\value{
A list including k centers, distance that every samples to k centers,
  clusters that samples belongs to
}
\description{
My own kmeans with raw R codes
}
\examples{
iris <- subset(iris, select = -Species)
KmeansR(dataset = iris, k = 3)
}
\seealso{
\code{\link{kmeans}}
}
