% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\docType{data}
\name{labels}
\alias{labels}
\title{Label matrix of MLKNN}
\description{
Label matrix of drug-indications (row: drug, column: indication),
  1 means drug interacts with indication, 0 means otherwise
}
\keyword{dataset}
