% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_roc.R
\name{multi_roc}
\alias{multi_roc}
\title{Plot multiple ROC curves in one graph}
\usage{
multi_roc(predictors, label)
}
\arguments{
\item{predictors}{Dataframe of your predicted possiblities of different models.}

\item{label}{Ideally, labels should be supplied as ordered factor(s), the lower level corresponding to the negative
class, the upper level to the positive class}
}
\value{
a list containing multiple ROCR objects
}
\description{
This function visualize multiple ROC curves and display values
of Area Under the Receiver Operating Characteristic curve (AUROC)
}
\seealso{
\code{\link{ROCR}}
}
