% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ols.R
\name{ols}
\alias{ols}
\title{Linear regression using ordinary least square}
\usage{
ols(y, x)
}
\arguments{
\item{y}{Response variable}

\item{x}{Independent variables}
}
\value{
A dataframe contains coefficients
}
\description{
Linear regression using ordinary least square
}
\examples{
library(tidyverse)
ols(y=diamonds$price,x=diamonds \%>\% select(-price) \%>\% select_if(is.numeric)) \%>\% print()
}
\seealso{
\code{\link{lm}}
}
