% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_walk.R
\name{random_walk}
\alias{random_walk}
\title{Using random walk algorithm to predict drug-drug interaction with restart}
\usage{
random_walk(
  similarity_matrix = matrix(),
  adjacent_matrix = matrix(),
  probability = 0.85
)
}
\arguments{
\item{similarity_matrix}{Similarity matrix of N drugs}

\item{adjacent_matrix}{Adjacent matrix of label (1 = known interaction, 0 = non-interaction)}

\item{probability}{The probability of jumping back ti initial point}
}
\value{
NxN Matrix of probability
}
\description{
Using random walk algorithm to predict drug-drug interaction with restart
}
\references{
Zhang el at (2017). Predicting potential drug-drug interactions by integrating chemical, biological, phenotypic and network data. BMC Bioinformatics 18:18
}
