% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sepcific_theme.R
\name{specific_theme}
\alias{specific_theme}
\title{Code to change the axis of ggplot image}
\usage{
specific_theme()
}
\value{
value of standard error
}
\description{
This function computes the standard error of the values in x. If na.rm is TRUE then missing values are
removed before computation proceeds.
}
\examples{
\dontrun{
library(tidyverse)
ggplot(iris, aes(x=Sepal.Length , y= Species, fill= Species))+geom_boxplot()+
specific_theme()
}
}
\seealso{
\code{\link{ggplot2}}
}
