% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_error.R
\name{standard_error}
\alias{standard_error}
\title{Standard Error}
\usage{
standard_error(x, na.rm = FALSE)
}
\arguments{
\item{x}{a numeric vector or an R object which is coercible to one by as.double(x)}

\item{na.rm}{logical. Should missing values be removed?}
}
\value{
value of standard error
}
\description{
This function computes the standard error of the values in x. If na.rm is TRUE then missing values are
removed before computation proceeds.
}
\examples{
standard_error(1:7)
}
\seealso{
\code{\link{sd}}
}
