% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcanoPlot.R
\name{volcanoPlot}
\alias{volcanoPlot}
\title{Output volcano plot with ggplot2}
\usage{
volcanoPlot(dt, mainTitle = "")
}
\arguments{
\item{dt}{Dataframe with specific columns}

\item{mainTitle}{Main title of the plot}
}
\value{
Object of ggplot2
}
\description{
Output volcano plot with ggplot2
}
\details{
Input data should be dataframe containing following columns:
log2FC: the value of fold change; padj: adjusted p value; threshold: Up-, down-regulated genes
}
